Refine documentation for waypoint sending
This was originally meant to add documentation for plane guided waypoints as a whole, but it was already added. This commit revises the content for this, and switches .send() documentation to be more explicit in regards to it being a generic command for any dialect.<MAVLink_Message_Type>.

Also adding a brief writeup for what each example is trying to achieve.